Update example with new build process
It is better for the build process to start with the linker, since it sets us
up for the future. It also offers a clean separation between the assembly and
build-time constraints, such as loading the binary at a particular address, or
setting the boot signature at the end of the boot sector.