use clang+lld for build process
No need to invoke the linker directly. Specify layout in the linker
script, and pass it to clang. This will make it easier to link C source
files in the future.

Cross-compilation toolchain was assembled, and captured in the
Dockerfile. Its image is hosted on hub.docker.com as
smasher164/oscons:booting.